Move seven-patch model to cruciform pincell script
